Extract file server service
