+++ imjoy/engine.py
@@ -6,6 +6,7 @@ from imjoy.helper import setup_logging
 from imjoy.options import parse_cmd_line
 from imjoy.env import bootstrap, prepare_env
 from imjoy.connection import create_connection_manager
+from imjoy.services import setup_services
 
 
 class Engine:
@@ -38,6 +39,7 @@ class Engine:
         self.conn = create_connection_manager(self)
         self.store = self.conn.store
         self.conn.setup()
+        setup_services(self)
 
     def start(self):
         """Start the engine."""
