+++ imjoy/services/__init__.py
@@ -1,1 +1,7 @@
 """Provide engine services."""
+from .file_server import setup_file_server
+
+
+def setup_services(engine):
+    """Set up engine services."""
+    setup_file_server(engine)
